4_get_bp/requirements.txt
pandas==1.1.5
kubeflow-katib
PyYAML==5.2
